+++ src/app/create/page.tsx
@@ -1,18 +1,15 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import ProductOptionsForm from '@/components/create/product-options-form';
 import Button from '@/components/design-system/button';
 import ButtonInput from '@/components/design-system/button-input';
-import usePostUrlCrawlingProduct from '@/hooks/mutations/use-post-url-crawling-product';
 import {
   type RegisteredBasketsParams,
   useRegisteredBaskets,
 } from '@/hooks/mutations/use-registered-baskets';
 import { LOCAL_STORAGE_KEY, Storage } from '@/libs/storage';
 import { useLoginModalStore } from '@/store/use-login-modal-store';
-import { useModalStore } from '@/store/use-modal-store';
-import { ERROR_CODE, ERROR_MESSAGES, type ErrorCode } from '@/utils/error-code';
 
 import * as styles from './index.css';
 
@@ -25,15 +22,6 @@ const CreatePage = () => {
   const [createForm, setCreateForm] = useState<CreateFormValue | null>(null);
   const accessToken = Storage.getItem(LOCAL_STORAGE_KEY.accessToken);
   const openLoginModal = useLoginModalStore(state => state.openLoginModal);
-  const onOpen = useModalStore(state => state.onOpen);
-
-  const {
-    mutate: fetchProduct,
-    data: product,
-    isPending,
-    isError,
-    error,
-  } = usePostUrlCrawlingProduct();
 
   const handleUrlInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setUrl(e.target.value);
@@ -41,7 +29,7 @@ const CreatePage = () => {
 
   const handleUrlInputButtonClick = () => {
     setShowProductsOptionsForm(true);
-    fetchProduct({ url });
+    // fetchProduct({ url });
   };
 
   const handleCreateFormSubmit = (e: React.FormEvent<HTMLFormElement>) => {
@@ -55,29 +43,6 @@ const CreatePage = () => {
     openLoginModal();
   }
 
-  const renderErrorMessage = (errorCode: ErrorCode) => {
-    const message = ERROR_MESSAGES[errorCode] || ERROR_MESSAGES[ERROR_CODE.INTERNAL_SERVER_ERROR];
-    const invalidUrl = error?.response?.data.error;
-
-    if (invalidUrl === 'UNSUPPORTED_URL') {
-      onOpen({
-        modalType: 'alert',
-        title: 'URL 주소를 다시 확인해주세요.',
-        description: '무신사 상품만 등록할 수 있어요.',
-      });
-    } else {
-      onOpen({ modalType: 'alert', title: message });
-    }
-  };
-
-  useEffect(() => {
-    if (isError) {
-      const errorCode =
-        ((error as Error)?.message as ErrorCode) || ERROR_CODE.INTERNAL_SERVER_ERROR;
-      renderErrorMessage(errorCode);
-    }
-  }, [isError, error]);
-
   return (
     <div className={styles.createPage}>
       <p className={styles.title}>지금 재입고 등록을 해보세요!👇</p>
@@ -90,13 +55,7 @@ const CreatePage = () => {
           label="URL 주소"
           buttonText="조회"
         />
-        {showProductsOptionsForm && (
-          <ProductOptionsForm
-            product={product}
-            isPending={isPending}
-            setCreateForm={setCreateForm}
-          />
-        )}
+        {showProductsOptionsForm && <ProductOptionsForm url={url} setCreateForm={setCreateForm} />}
         <Button
           style={{ width: '100%', cursor: 'pointer' }}
           disabled={!createForm?.firstOption}
