+++ src/components/create/product-options-form/index.tsx
@@ -1,28 +1,48 @@
 import { type Dispatch, type SetStateAction, useEffect } from 'react';
 import Image from 'next/image';
 import Select from '@/components/design-system/select';
-import { type UsePostUrlCrawlingProductResponse } from '@/hooks/mutations/use-post-url-crawling-product';
 import { type RegisteredBasketsParams } from '@/hooks/mutations/use-registered-baskets';
+import { useGetProductsCrawling } from '@/hooks/queries/use-get-products-crawling';
+import { useModalStore } from '@/store/use-modal-store';
+import { ERROR_CODE, ERROR_MESSAGES, type ErrorCode } from '@/utils/error-code';
+import type { AxiosError } from 'axios';
 
 import * as styles from './index.css';
 
 interface ProductOptionsFormProps {
-  product?: UsePostUrlCrawlingProductResponse;
-  isPending: boolean;
+  url: string;
   setCreateForm: Dispatch<SetStateAction<RegisteredBasketsParams | null>>;
 }
 
-const ProductOptionsForm = ({ product, isPending, setCreateForm }: ProductOptionsFormProps) => {
+const ProductOptionsForm = ({ url, setCreateForm }: ProductOptionsFormProps) => {
+  const onOpen = useModalStore(state => state.onOpen);
+
+  const { data: product, isPending, isError, error } = useGetProductsCrawling(url);
+
   useEffect(() => {
     if (product) {
+      const {
+        firstOptions: _firstOptions,
+        secondOptions: _secondOptions,
+        ...restProduct
+      } = product;
       setCreateForm({
-        ...product,
+        ...restProduct,
+        imageUrlList: [product.thumbnailUrl],
         firstOption: product.firstOptions[0] ?? '',
         secondOption: product.secondOptions[0] ?? null,
         thirdOption: product.thirdOptions[0] ?? null,
       });
     }
   }, [product, setCreateForm]);
+
+  useEffect(() => {
+    if (isError) {
+      const errorCode =
+        ((error as Error)?.message as ErrorCode) || ERROR_CODE.INTERNAL_SERVER_ERROR;
+      renderErrorMessage(errorCode);
+    }
+  }, [isError, error]);
 
   const handleOptionsChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     const { name, value } = e.target;
@@ -33,10 +53,25 @@ const ProductOptionsForm = ({ product, isPending, setCreateForm }: ProductOption
 
   if (!product) return null;
 
+  const renderErrorMessage = (errorCode: ErrorCode) => {
+    const message = ERROR_MESSAGES[errorCode] || ERROR_MESSAGES[ERROR_CODE.INTERNAL_SERVER_ERROR];
+    const invalidUrl = (error as AxiosError)?.response?.data;
+
+    if (invalidUrl === 'UNSUPPORTED_URL') {
+      onOpen({
+        modalType: 'alert',
+        title: 'URL 주소를 다시 확인해주세요.',
+        description: '무신사 상품만 등록할 수 있어요.',
+      });
+    } else {
+      onOpen({ modalType: 'alert', title: message });
+    }
+  };
+
   return (
     <>
       <Image
-        src={product.thumbnailUrl}
+        src={`https://image.msscdn.net${product.thumbnailUrl}`}
         alt="Product Preview"
         className={styles.previewImage}
         width={800}
