+++ src/hooks/mutations/use-registered-baskets.ts
@@ -12,7 +12,6 @@ export interface RegisteredBasketsParams {
   name: string;
   brand: string;
   imageUrlList: string[];
-  category: string;
   firstCategory: string;
   secondCategory: string;
   price: number;
